integrating djangorestframework and handlebars.js for json handshakes between the server and the client
the default functionality of DRF is in place, the integration of handlebars is tried out but not completed; it remains to secure the API for user access (permissions in views); in case of handlebars.js it remains to rewrite some of the existing django templates

handlebars conflict with django's template syntax; the verbose tag for django templates will not arrive before the version 1.5, this means that you either have to precompile the templates (tempting solution, faster mobile load times, though difficult in development) or fetch the templates via a http-request. For now I have chosen the latter strategy.